deposit dialog polish
